backend/requirements.txt
Django>=4.1,<4.2
wagtail>=4.1,<4.2
gunicorn>=20.1.0

python-dotenv
graphene-django
django-cors-headers
django-sslserver